requirements.txt
python-telegram-bot==20.3
transformers==4.34.0
torch
PyMySQL
python-dotenv
#py -3.11 -m venv ..venv\Scripts\activate
#pip install -r requirements.txt
#pip install accelerate
#pip install transformers[torch]
